+++ CHANGELOG.md
@@ -36,7 +36,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
-## [v0.13.0] - 2022-04-19
+## [v0.14.0] - 2022-04-19
 
 ### API Breaking